Extract methods that manipulate the button
So far changing the text and the action listener on the button was done inline
in the `CheckStateTask`.

Since this code will be reused by the `UnmountTask` and `MountTask` to toggle
the state of the button after they are done, this commit extracts the required
functionality into its own methods.