Extract tasks into individual classes
So far the asynchronous tasks were private inner classes of the `MainActivity`,
which was technically attractive, since they were well encapsulated, but made
for a rather large and nested `MainActivity`.

Since this is a small app that will ever only consist of one activity that
manages all the functionality, I decided to expose the private classes in favor
of enhanced code structure.